Simplify build tooling: use tsc and node:test
- Replace tsup with tsc for TypeScript compilation (ESM-only)
- Replace vitest with node:test (zero runtime dependencies)
- Remove .nvmrc (package.json engines is sufficient)
- Update CI workflow for simplified tooling
- Convert tests from TypeScript to JavaScript (node:test)
- Update engines to Node.js >=20.0.0

Closes #17
