+++ .github/workflows/ci.yml
@@ -31,11 +31,8 @@ jobs:
             - name: Lint
               run: npm run lint
 
-            - name: Type check
-              run: npm run typecheck
-
-            - name: Test
-              run: npm run test:coverage
-
             - name: Build
               run: npm run build
+
+            - name: Test
+              run: npm test
