+++ package.json
@@ -5,18 +5,11 @@
     "description": "EMV / Chip and PIN library for PC/SC card readers",
     "exports": {
         ".": {
-            "import": {
-                "types": "./dist/index.d.ts",
-                "default": "./dist/index.js"
-            },
-            "require": {
-                "types": "./dist/index.d.cts",
-                "default": "./dist/index.cjs"
-            }
+            "types": "./dist/index.d.ts",
+            "default": "./dist/index.js"
         }
     },
-    "main": "./dist/index.cjs",
-    "module": "./dist/index.js",
+    "main": "./dist/index.js",
     "types": "./dist/index.d.ts",
     "files": [
         "dist"
@@ -25,17 +18,15 @@
         "node": ">=20.0.0"
     },
     "scripts": {
-        "build": "tsup",
-        "dev": "tsup --watch",
-        "typecheck": "tsc --noEmit",
-        "lint": "eslint src/ tests/",
-        "lint:fix": "eslint src/ tests/ --fix",
+        "build": "tsc",
+        "clean": "rm -rf dist",
+        "lint": "eslint src/",
+        "lint:fix": "eslint src/ --fix",
         "format": "prettier --write .",
         "format:check": "prettier --check .",
-        "test": "vitest run",
-        "test:watch": "vitest",
-        "test:coverage": "vitest run --coverage",
-        "prepublishOnly": "npm run lint && npm run typecheck && npm run test && npm run build"
+        "test": "node --test tests/*.test.js",
+        "pretest": "npm run build",
+        "prepublishOnly": "npm run clean && npm run lint && npm run build && npm run test"
     },
     "author": "tomkp <tom@tomkp.com>",
     "keywords": [
@@ -64,13 +55,10 @@
     "devDependencies": {
         "@eslint/js": "^9.17.0",
         "@types/node": "^22.0.0",
-        "@vitest/coverage-v8": "^4.0.16",
         "eslint": "^9.17.0",
         "globals": "^15.14.0",
         "prettier": "^3.4.2",
-        "tsup": "^8.5.1",
-        "typescript": "^5.9.3",
-        "typescript-eslint": "^8.50.1",
-        "vitest": "^4.0.16"
+        "typescript": "^5.7.0",
+        "typescript-eslint": "^8.18.0"
     }
 }
