--- vitest.config.ts
@@ -1,19 +1,0 @@
-import { defineConfig } from 'vitest/config';
-
-export default defineConfig({
-    test: {
-        globals: true,
-        coverage: {
-            provider: 'v8',
-            reporter: ['text', 'lcov', 'html'],
-            include: ['src/**/*.ts'],
-            exclude: ['src/**/*.d.ts', 'src/types.ts', 'src/index.ts'],
-            thresholds: {
-                statements: 80,
-                branches: 80,
-                functions: 80,
-                lines: 80,
-            },
-        },
-    },
-});
